+++ RELEASES.md
@@ -18,7 +18,7 @@
 - Added `Direction::try_new()` to fallibly create a new `Direction` struct (which cannot be created from the zero vector).
 - Rotation is now measured from the positive x axis in a counterclockwise direction. This applies to both `Rotation` and `Direction`.
   - This increases consistency with `glam` and makes trigonometry easier.
-- Added `Direction::try_new` which never panics.
+- Added `Direction::try_from` which never panics; consider using this in place of `Direction::new`.
 - Converting from a `Direction` (which uses a `Vec2` of `f32`'s internally) to a `Rotation` (which uses exact decidegrees) now has special cases to ensure all eight cardinal directions result in exact degrees.
   - For example, a unit vector pointing to the Northeast now always converts to a `Direction` with exactly 1350 decidegrees.
   - Rounding errors may still occur when converting from arbitrary directions to the other 3592 discrete decidegrees.